% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dept.R
\name{deleteAllcache_dept}
\alias{deleteAllcache_dept}
\title{规则表更新}
\usage{
deleteAllcache_dept(token = "9ADDE293-1DC6-4EBC-B8A7-1E5CC26C1F6C")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
规则表更新
}
\examples{
deleteAllcache_dept()
}
