% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workcenter.R
\name{deleteCache_workcenter}
\alias{deleteCache_workcenter}
\title{核算维度_责任中心更新}
\usage{
deleteCache_workcenter(token = "9ADDE293-1DC6-4EBC-B8A7-1E5CC26C1F6C")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
核算维度_责任中心更新
}
\examples{
deleteCache_workcenter()
}
