% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosucompany.R
\name{insertCache_cosucompany}
\alias{insertCache_cosucompany}
\title{科目更新}
\usage{
insertCache_cosucompany(token = "9ADDE293-1DC6-4EBC-B8A7-1E5CC26C1F6C")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
科目更新
}
\examples{
insertCache_cosucompany()
}
