% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rditem.R
\name{insertCache_rditem}
\alias{insertCache_rditem}
\title{研发项目更新}
\usage{
insertCache_rditem(token = "9ADDE293-1DC6-4EBC-B8A7-1E5CC26C1F6C")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
研发项目更新
}
\examples{
insertCache_rditem()
}
