% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctreclass.R
\name{ViewAcctreclass}
\alias{ViewAcctreclass}
\title{核算维度_重分类查询}
\usage{
ViewAcctreclass(token = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
核算维度_重分类查询
}
\examples{
ViewAcctreclass()
}
