% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosucompany.R
\name{ViewCosucompany}
\alias{ViewCosucompany}
\title{往来单位查询}
\usage{
ViewCosucompany(token = "9ADDE293-1DC6-4EBC-B8A7-1E5CC26C1F6C")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
往来单位查询
}
\examples{
ViewCosucompany()
}
