% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rditem.R
\name{ViewRditem}
\alias{ViewRditem}
\title{研发项目查询}
\usage{
ViewRditem(token = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
研发项目查询
}
\examples{
ViewRditem()
}
