% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workcenter.R
\name{ViewWorkcenter}
\alias{ViewWorkcenter}
\title{责任中心查询}
\usage{
ViewWorkcenter(token = "9ADDE293-1DC6-4EBC-B8A7-1E5CC26C1F6C")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
责任中心查询
}
\examples{
ViewWorkcenter()
}
