% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acct.R, R/cosucompany.R
\name{deleteCache_acct}
\alias{deleteCache_acct}
\title{科目更新}
\usage{
deleteCache_acct(token = "057A7F0E-F187-4975-8873-AF71666429AB")

deleteCache_acct(token = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{token}{}
}
\value{
无返回值

无返回值
}
\description{
科目更新

科目更新
}
\examples{
deleteCache_acct()
deleteCache_acct()
}
