% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosucompany.R
\name{insertCache_cosucompany}
\alias{insertCache_cosucompany}
\title{科目更新}
\usage{
insertCache_cosucompany(token = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
科目更新
}
\examples{
insertCache_cosucompany()
}
