% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acctreclass.R
\name{ViewAcctreclass}
\alias{ViewAcctreclass}
\title{核算维度_重分类查询}
\usage{
ViewAcctreclass(token = "9ADDE293-1DC6-4EBC-B8A7-1E5CC26C1F6C")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
核算维度_重分类查询
}
\examples{
ViewAcctreclass()
}
