% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosucompany.R
\name{ViewCosucompany}
\alias{ViewCosucompany}
\title{往来单位查询}
\usage{
ViewCosucompany(token = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
往来单位查询
}
\examples{
ViewCosucompany()
}
