% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workcenter.R
\name{ViewWorkcenter}
\alias{ViewWorkcenter}
\title{责任中心查询}
\usage{
ViewWorkcenter(token = "057A7F0E-F187-4975-8873-AF71666429AB")
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
责任中心查询
}
\examples{
ViewWorkcenter()
}
